requirements.txt
pytest == 7.2.0
selene == 2.0.0b16
voluptuous~=0.13.1
pytest-voluptuous
requests~=2.28.2
allure-pytest
python-dotenv~=1.0.0

curlify~=2.2.1
Faker~=17.6.0